python/pyproject.toml
[tool.poetry]
name = "my-app"
version = "0.1.0"
description = ""
authors = ["Mihai Ionut Vilcu <ionutvmi@gmail.com>"]
readme = "README.md"
packages = [{ include = "my_app", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
poethepoet = "^0.20.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-watch = "^4.2.0"
pytest-testmon = "^2.0.8"
colorama = "^0.4.6"
watchdog = "^3.0.0"
pytest-dotenv = "^0.5.2"

[tool.pytest.ini_options]
addopts = '--strict-markers'
markers = ["fast", "furious"]

# Tasks/Scripts
[tool.poe.tasks.start]
help = "Start the application"
cmd = "python src/my_app/main.py"

[tool.poe.tasks.dev]
help = "Start the application in watch mode"
cmd = "python pymon.py --all src/my_app/main.py"

[tool.poe.tasks.test]
help = "Run all the tests"
cmd = "pytest"

[tool.poe.tasks.test_watch]
help = "Run all tests in the watch mode"
cmd = "ptw -- --testmon"

[tool.poe.tasks.test_fast]
help = "Run all the tests decorated with @pytest.mark.fast"
cmd = "pytest -m fast"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
